fix travis

+++ publish-javadoc.sh
@@ -1,13 +1,16 @@
-# Get to the Travis build directory, configure git and clone the repo
-cd $HOME
-git config --global user.email "travis@travis-ci.org"
-git config --global user.name "travis-ci"
-git clone --quiet --branch=gh-pages https://${GH_TOKEN}@github.com/deichbrise/binf gh-pages > /dev/null
 
-# Commit and Push the Changes
-cd gh-pages
-git rm -rf ./javadoc-latest
-cp -Rf $HOME/javadoc ./javadoc-latest
-git add -f .
-git commit -m "Lastest javadoc on successful travis build $TRAVIS_BUILD_NUMBER auto-pushed to gh-pages"
-git push -fq origin gh-pages > /dev/null
+if [ "$TRAVIS_REPO_SLUG" == "deichbrise/binf" ] && [ "$TRAVIS_JDK_VERSION" == "oraclejdk8" ] && [ "$TRAVIS_PULL_REQUEST" == "false" ] && [ "$TRAVIS_BRANCH" == "master" ]; then
+    # Get to the Travis build directory, configure git and clone the repo
+    cd $HOME
+    git config --global user.email "travis@travis-ci.org"
+    git config --global user.name "travis-ci"
+    git clone --quiet --branch=gh-pages https://${GH_TOKEN}@github.com/deichbrise/binf gh-pages > /dev/null
+
+    # Commit and Push the Changes
+    cd gh-pages
+    git rm -rf ./javadoc-latest
+    cp -Rf $HOME/javadoc ./javadoc-latest
+    git add -f .
+    git commit -m "Lastest javadoc on successful travis build $TRAVIS_BUILD_NUMBER auto-pushed to gh-pages"
+    git push -fq origin gh-pages > /dev/null
+fi
